requirements.txt
requests==2.27.1
python-dotenv==0.20.0
python-telegram-bot==13.13
